requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
httpx
python-jose
passlib[bcrypt]
bcrypt
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
